refactor: Apply common modal component (#62)
