--- src/pages/listPage/MyCredit/components/Modal.module.css
@@ -1,34 +1,0 @@
-.background {
-  position: fixed;
-  inset: 0;
-  z-index: 1;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  background-color: #02000ecc;
-}
-
-.container {
-  padding: 24px 16px;
-  border-radius: 8px;
-  background-color: var(--black-100);
-  color: var(--white-100);
-}
-
-.head {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 24px;
-}
-
-.title {
-  font-size: 18px;
-  font-weight: 500;
-}
-
-.button {
-  width: 24px;
-  height: 24px;
-  background-image: url('@/assets/icons/close-modal.svg');
-}
