+++ src/pages/listPage/MyCredit/components/TopUpModalContent.module.css
@@ -3,6 +3,7 @@
   flex-direction: column;
   gap: 8px;
   width: 296px;
+  margin-top: 24px;
 }
 
 .radio {
